move watcher functionality to core util.rs + read db users and populate app state users list is now in core util.rs
